inclusion of query manager
Conflicts:
	analysis/common/functions.php
	analysis/index.php
	capture/stream/onepercent.php
	capture/stream/track.php

+++ capture/stream/controller.php
@@ -10,48 +10,51 @@ include "../common/functions.php";           // load capture function file
 
 $dbh = pdo_connect();
 create_error_logs($dbh);
-//$dbh = null;                                 // https://bugs.php.net/bug.php?id=62065
 
 $roles = unserialize(CAPTUREROLES);
-
-// first handle all instruction sent by the webinterface to the controller (ie. the instruction queue)
 
 $commands = array();
 foreach ($roles as $role) {
     $commands[$role] = array();
 }
 
-$sql = "select task, instruction from tcat_controller_tasklist order by id asc lock in share mode";
-foreach ($dbh->query($sql) as $row) {
-    if (!array_key_exists($row['task'], $commands)) {
+// first handle all instruction sent by the webinterface to the controller (ie. the instruction queue)
+$rec = $dbh->prepare("SHOW TABLES LIKE 'tcat_controller_tasklist'");
+if ($rec->execute() && $rec->rowCount() > 0) {
+    $sql = "select task, instruction from tcat_controller_tasklist order by id asc lock in share mode";
+    foreach ($dbh->query($sql) as $row) {
+        if (!array_key_exists($row['task'], $commands)) {
+            continue;
+        }
+        $commands[$row['task']][] = $row;
+    }
+
+    // do not leave any unknown tasks linger
+    $sql = 'truncate tcat_controller_tasklist';
+    $h = $dbh->prepare($sql);
+    $res = $h->execute();
+}
+
+foreach ($roles as $role) {
+
+    if (!empty($commands[$role])) {
+        foreach ($commands[$role] as $command) {
+            logit("controller.log", "received instruction to execute '" . $command['instruction'] . "' for script $role");
+            switch ($command['instruction']) {
+                case "reload": {
+                        // reload configuration for a task
+                        logit("controller.log", $command['instruction'] . " $role");
+                        controller_reload_config_role($role);
+                        break;
+                    }
+                default: {
+                        break;
+                    }
+            }
+        }
         continue;
     }
-    $commands[$row['task']][] = $row;
-}
-
-// do not leave any unknown tasks linger
-$sql = 'truncate tcat_controller_tasklist';
-$h = $dbh->prepare($sql);
-$res = $h->execute();
-
-foreach ($roles as $role) {
-    foreach ($commands[$role] as $command) {
-        logit("controller.log", "received instruction to execute '" . $command['instruction'] . "' for script $role");
-        switch ($command['instruction']) {
-            case "reload": {
-                    // reload configuration for a task
-                    controller_reload_config_role($role);
-                    break;
-                }
-            default: {
-                    break;
-                }
-        }
-    }
-}
-
-foreach ($roles as $role) {
-
+    
     if (defined('IDLETIME')) {
         $idletime = IDLETIME;
     } else {
@@ -79,6 +82,7 @@ foreach ($roles as $role) {
         if ($out == 2)
             $running = true;
 
+
         // check whether the process has been idle for too long
         logit("controller.log", "script $role called - pid:" . $pid . "  idle:" . (time() - $last));
 
@@ -89,7 +93,6 @@ foreach ($roles as $role) {
 
             if ($running) {
 
-                //posix_kill(SIGTERM, $pid);
                 $res = exec("kill -s SIGTERM $pid");
 
                 logit("controller.log", "script $role was idle for more than " . $idletime . " seconds - killing and starting");
@@ -97,7 +100,7 @@ foreach ($roles as $role) {
 
                 sleep(6);       // we need some time to allow graceful exit
 
-                passthru("php " . BASE_FILE . "capture/stream/$role.php > /dev/null 2>&1 &");
+                passthru(PHP_CLI . " " . BASE_FILE . "capture/stream/$role.php > /dev/null 2>&1 &");
             }
         }
     }
@@ -105,15 +108,7 @@ foreach ($roles as $role) {
 
         logit("controller.log", "script $role was not running - starting");
 
-        passthru("php " . BASE_FILE . "capture/stream/$role.php > /dev/null 2>&1 &");
+        passthru(PHP_CLI . " " . BASE_FILE . "capture/stream/$role.php > /dev/null 2>&1 &");
     }
 }
-
-function logit($file, $message) {
-
-    $file = BASE_FILE . "logs/" . $file;
-    $message = date("Y-m-d H:i:s") . " " . $message . "\n";
-    file_put_contents($file, $message, FILE_APPEND);
-}
-
 ?>
